Add random methods to header. Init TitleScreen
