+++ sfml-pong/include/GameScreen.h
@@ -4,19 +4,9 @@
 #include "Paddle.h"
 #include "Text.h"
 #include "HalfCourtLine.h"
-#include <random>
 
 namespace Pong
 {
-	inline int random(int min, int max)
-	{
-		static std::default_random_engine random_engine;
-		random_engine.seed(std::random_device{}());
-		return std::uniform_int_distribution<>{min, max}(random_engine);
-	}
-
-	inline bool coin_toss() { return random(0, 1); }
-
 	class GameScreen : public Screen
 	{
 	public:
@@ -39,12 +29,11 @@ namespace Pong
 
 		virtual void serve();
 
+		void draw(sf::RenderTarget* target) override;
 		void on_start() override;
 		void update(float dt) override;
-		void draw(sf::RenderTarget* target) override;
 		void handle_input(sf::Event event) override;
-
-		void on_stop() override {}
+		void on_stop() override;
 
 	private:
 		const int k_serve_delay;
